Also introduce Jenssegers\Mongodb\Eloquent\Model class for 2.2
