+++ README.md
@@ -106,7 +106,7 @@ Eloquent
 This package includes a MongoDB enabled Eloquent class that you can use to define models for corresponding collections.
 
 ```php
-use Jenssegers\Mongodb\Model as Eloquent;
+use Jenssegers\Mongodb\Eloquent\Model as Eloquent;
 
 class User extends Eloquent {}
 ```
@@ -114,7 +114,7 @@ class User extends Eloquent {}
 Note that we did not tell Eloquent which collection to use for the `User` model. Just like the original Eloquent, the lower-case, plural name of the class will be used as the table name unless another name is explicitly specified. You may specify a custom collection (alias for table) by defining a `collection` property on your model:
 
 ```php
-use Jenssegers\Mongodb\Model as Eloquent;
+use Jenssegers\Mongodb\Eloquent\Model as Eloquent;
 
 class User extends Eloquent {
 
@@ -126,7 +126,7 @@ class User extends Eloquent {
 **NOTE:** Eloquent will also assume that each collection has a primary key column named id. You may define a `primaryKey` property to override this convention. Likewise, you may define a `connection` property to override the name of the database connection that should be used when utilizing the model.
 
 ```php
-use Jenssegers\Mongodb\Model as Eloquent;
+use Jenssegers\Mongodb\Eloquent\Model as Eloquent;
 
 class MyModel extends Eloquent {
 
@@ -142,7 +142,7 @@ Everything else works just like the original Eloquent model. Read more about the
 You may also register an alias for the MongoDB model by adding the following to the alias array in `app/config/app.php`:
 
 ```php
-'Moloquent'       => 'Jenssegers\Mongodb\Model',
+'Moloquent'       => 'Jenssegers\Mongodb\Eloquent\Model',
 ```
 
 This will allow you to use the registered alias like:
@@ -232,10 +232,10 @@ MongoDB Support => enabled
 
 #### Argument 2 passed to Illuminate\Database\Query\Builder::__construct() must be an instance of Illuminate\Database\Query\Grammars\Grammar, null given
 
-To solve this, you will need to check two things. First check if your model is extending the correct class; this class should be `Jenssegers\Mongodb\Model`. Secondly, check if your model is using a MongoDB connection. If you did not change the default database connection in your database configuration file, you need to specify the MongoDB enabled connection. This is what your class should look like if you did not set up an alias and change the default database connection:
+To solve this, you will need to check two things. First check if your model is extending the correct class; this class should be `Jenssegers\Mongodb\Eloquent\Model`. Secondly, check if your model is using a MongoDB connection. If you did not change the default database connection in your database configuration file, you need to specify the MongoDB enabled connection. This is what your class should look like if you did not set up an alias and change the default database connection:
 
 ```php
-use Jenssegers\Mongodb\Model as Eloquent;
+use Jenssegers\Mongodb\Eloquent\Model as Eloquent;
 
 class User extends Eloquent {
 
@@ -530,7 +530,7 @@ Eloquent allows you to work with Carbon/DateTime objects instead of MongoDate ob
 Example:
 
 ```php
-use Jenssegers\Mongodb\Model as Eloquent;
+use Jenssegers\Mongodb\Eloquent\Model as Eloquent;
 
 class User extends Eloquent {
 
@@ -557,7 +557,7 @@ Supported relations are:
 Example:
 
 ```php
-use Jenssegers\Mongodb\Model as Eloquent;
+use Jenssegers\Mongodb\Eloquent\Model as Eloquent;
 
 class User extends Eloquent {
 
@@ -572,7 +572,7 @@ class User extends Eloquent {
 And the inverse relation:
 
 ```php
-use Jenssegers\Mongodb\Model as Eloquent;
+use Jenssegers\Mongodb\Eloquent\Model as Eloquent;
 
 class Item extends Eloquent {
 
@@ -587,7 +587,7 @@ class Item extends Eloquent {
 The belongsToMany relation will not use a pivot "table", but will push id's to a __related_ids__ attribute instead. This makes the second parameter for the belongsToMany method useless. If you want to define custom keys for your relation, set it to `null`:
 
 ```php
-use Jenssegers\Mongodb\Model as Eloquent;
+use Jenssegers\Mongodb\Eloquent\Model as Eloquent;
 
 class User extends Eloquent {
 
@@ -607,7 +607,7 @@ Other relations are not yet supported, but may be added in the future. Read more
 If you want to embed models, rather than referencing them, you can use the `embedsMany` relation. This relation is similar to the `hasMany` relation, but embeds the models inside the parent object.
 
 ```php
-use Jenssegers\Mongodb\Model as Eloquent;
+use Jenssegers\Mongodb\Eloquent\Model as Eloquent;
 
 class User extends Eloquent {
 
@@ -702,7 +702,7 @@ $books = $user->books()->where('rating', '>', 5)->orderBy('title')->get();
 The embedsOne relation is similar to the EmbedsMany relation, but only embeds a single model.
 
 ```php
-use Jenssegers\Mongodb\Model as Eloquent;
+use Jenssegers\Mongodb\Eloquent\Model as Eloquent;
 
 class Book extends Eloquent {
 
@@ -772,7 +772,7 @@ class User extends Eloquent {
 And the Mongodb-based Message model:
 
 ```php
-use Jenssegers\Mongodb\Model as Eloquent;
+use Jenssegers\Mongodb\Eloquent\Model as Eloquent;
 
 class Message extends Eloquent {
