+++ tests/models/Address.php
@@ -1,6 +1,6 @@
 <?php
 
-use Jenssegers\Mongodb\Model as Eloquent;
+use Jenssegers\Mongodb\Eloquent\Model as Eloquent;
 
 class Address extends Eloquent {
