+++ tests/models/Client.php
@@ -1,6 +1,6 @@
 <?php
 
-use Jenssegers\Mongodb\Model as Eloquent;
+use Jenssegers\Mongodb\Eloquent\Model as Eloquent;
 
 class Client extends Eloquent {
